Split subagents into separate files for development
	new file:   AbstractAgents/CameraSubAgent.py
	new file:   AbstractAgents/CcdCoolerSubAgent.py
	new file:   AbstractAgents/DomeSubAgent.py
	new file:   AbstractAgents/FilterWheelSubAgent.py
	new file:   AbstractAgents/MountSubAgent.py
	modified:   AbstractAgents/SubAgent.py
